Feature: Commented time.Sleep in worker for concurrent simulation
time.Sleep can be used to make the concurrent simulation more
deterministic by forcing the goroutines to wait if there is nothing
on the receive channel. This can be tweaked to not incur too much of a
performance penalty but the depth also has to be set such that it is not
too high over the actives of the deterministic answer or the simulation
will run much longer because each extra depth cycle will trigger the
wait time while all the actives that could ever be discovered will have
been found.