Feature: Fixed concurrent algorithm to timeout when is possible
The select statement takes a node from the receive channel if there
is a value to be received. If there is none, the select statement
blocks until either the timeout channel sends a value and terminates
the goroutine or it receives a value from the receive channel. When
it receives a value on the receive channel the timeout channel is
reset. Timeout is currently 10 Milliseconds, which introduces a
lower bound to the processing if more processing resources
(goroutines or depth) is allocated to the simulation.